Remove antd dependencies from create API using WSDL and other UI improvments
