+++ features/apimgt/org.wso2.carbon.apimgt.publisher.feature/src/main/resources/publisher/source/src/app/components/Base/Footer/Footer.js
@@ -20,9 +20,11 @@ import React from 'react'
 
 export const Footer = () => (
     <footer className="footer">
-        <div style={{margin: "5% 0 0 0", 'textAlign': "center"}} className="container-fluid">
+        <div style={{textAlign: "center"}} className="container-fluid">
             <p>WSO2 APIM Publisher v3.0.0 | © 2017 <a href="http://wso2.com/" target="_blank">
-                <img className="footer-logo" src="/publisher/public/app/images/logo-footer.svg" alt="wso2-logo"/> Inc</a>.</p>
+                <img className="footer-logo" src="/publisher/public/app/images/logo-footer.svg" alt="wso2-logo"/>
+                 Inc</a>.
+            </p>
         </div>
     </footer>
 );
